Update search fragment

+++ app/src/main/java/com/example/zotfit/DailyData.kt
@@ -1,3 +1,3 @@
 package com.example.zotfit
 
-public data class DailyData(val username: String, val calorories: Int, val fat: Int, val protein: Int, val carbohydrates:Int)
+public data class DailyData(val username: String, val calorories: Float, val fat: Float, val protein: Float, val carbohydrates:Float)
